Fixed Multiple Duplicate pk injections (#9)
Added calls with SecretToken for AccountService
Fixed multiple Methods supposed to return an IEnumerable<Type> instead of plain Type

+++ IEXSharp/Model/Shared/Response/Trade.cs
@@ -4,7 +4,7 @@
 	{
 		public decimal price { get; set; }
 		public int size { get; set; }
-		public int tradeId { get; set; }
+		public long tradeId { get; set; }
 		public bool isISO { get; set; }
 		public bool isOddLot { get; set; }
 		public bool isOutsideRegularHours { get; set; }
